deixa rules em arquivo único
