--- src/regexes.js
@@ -1,9 +1,0 @@
-/*
- * Acentos
- */
-exports.regexAccents = /[áâéêôóí]/
-
-/*
- * Vogais
- */
-exports.regexVowels = /[aeiou]/
