--- src/rules/x.js
@@ -1,20 +1,0 @@
-/*
- * Pluralizar substantivos terminados em -x.
- *
- * Regras básicas:
- *
- * Não faz alterações.
- */
-
-const suffixes = [
-  'x'
-]
-
-function handler(word) {
-  return word
-}
-
-module.exports = {
-  suffixes,
-  handler
-}
